add optnone to avoid memset idiom recognition
In the steady state loop clang used to recognize a memset and it transformed
that loop into a call to memset. After outlining the hot loop to be able to add
__attribute__((optnone)) clang stopped loop idiom recognizing it, and we get a
speedup.

On aarch64, the chromium benchmark pdf was inflating by up to 10% slower and all
other benchmarks were not impacted much:
https://gist.github.com/sebpop/0b53c62dea18bb853184f4318a65f1e4

On x86_64 this change consistently leads to regressions:
https://gist.github.com/sebpop/c58f665b80fb7dff1ed2ed029aec674f